Configure firewall and balancer states lifetime options
The commit introduces two groups of state timeout options
The first one options group is for stateful firewall:
 - stateful_firewall_tcp_syn_timeout
 - stateful_firewall_tcp_syn_ack_timeout
 - stateful_firewall_tcp_fun_timeout
This options define the lifetime for TCP session depending on the last
processed session packet.

And the second group is for balancer states:
 - balancer_tcp_timeout
 - balancer_tcp_syn_timeout
 - balancer_tcp_syn_ack_timeout
 - balancer_tcp_fin_timeout
 - balancer_udp_timeout
 - balancer_other_protocols_timeout
All of them are defining timeouts for balancer sessions instead of
`balancer_state_ttl` which is obsoleted now.

Closes #61