Control plane workers config parsing
Adds parsing of control plane workers configuration from config file.
Expected config file section is:

"controlPlaneWorkers": [
    {
        "core" : 42,
        "interfaces": [ <interface_name_string> ]
    }
]

In case section is not present, configuration for single control plane
worker, responsible for all dataplane interfaces and residing on
controlPlaneCoreId is generated. This Is done to ease transition on
already deployed appliances.

Section is checked for
- Every configured dataplane interface is assigned to a control plane worker
- Each interface is assigned to single control plane worker
- Interfaces not configured for use by dataplane, but listed in
  controlplane workers are treated as errors