Turn static is_nonterm_action to a rule_t method is_term
First of all, is_nonterm makes implementation a bit more complex, as
everywhere we use !is_nonterm, which does not makes sense. It would be
better to just use is_term instead.

Next, we don't need this as a static function defined in acl.cpp, as
acl_compiler.cpp also wants the same functionality to fill
`rule.terminating` field. So we make this a rule_t method.