Add new check-state action in unwind step
Conveniently, parser tokens were already present in the codebase, so we don't
have to change them. Although we discussed that we might want to enrich our
ipfw syntax by allowing the check-state action to go with filters like
`add check-state tcp from X to Y`, this is a task for a future improvement.

For now, check-state rules do not hold any information aside from the fact
that a check-state should occur. Therefore, the related structure does not
have any fields and is effectively a minimal one-byte struct.
Consequently, we need some value to uniquely identify and hash it.