Add new Actions class to replace old value_t class
The value_t class previously represented a sequence of actions to be
performed in the dataplane upon packet match. It contained a list of
dump_ids and a flow (allow/deny/etc.), which was awkward as not all
fields were necessary if certain rule types weren't matched.
Additionally, adding new classes for new rule types was inconvenient.

The new Actions class addresses these issues by storing a vector of
"abstract" Actions that are matched and executed in the dataplane one
by one, as intended. This makes the compilation of total_table
(mapping from group_id to the actions to be performed for the packet)
readable and efficient.