Use Actions instead of old value_t to represent actions for the packet
The previous implementation used a `value_t` structure to represent
actions required for packet processing. This approach was cumbersome
when adding new functionality, as it required modifications
to a monolithic structure.

Changes:
- Replaced `value_t` with a more flexible `Actions` structure.
- Removed the outdated `value_t` definition and its associated methods.
- Simplified the algorithm for filling {group_id, acl_id} by directly iterating
over groups and not through each rule_id (time complexity has not changed,
as we still need to iterate over all rule_ids, but we do that on "prepare" step now)

Actions are now managed in a more modular and decoupled manner.