Rename common::acl::action_t to dump_t
This structure lost it's meaning as one struct to store all
non-terminating rule data. This is not a good thing anyway
as tHis will pollute the abstraction and the class size will
grow when we will introduce new nonterminating
kind of rules. Defining a new struct/class for each rule kind and
then using it in an std::variant is a preffered approach.
Therefore this class should not be called "action_t" but rather "dump_t"