Update firewall rules in autotests to include check-state
Previously if the packet should be dropped, the state was being checked and if
it was found the packet would not be dropped.
This behavior was incorrect according to the ipfw-like semantics.

Now, the behavior has been updated to be ipfw-like, meaning state checking is
not performed unless a check-state rule is explicitly matched.