Add autotests for keep-state firewall rules in existing test suite
Previously, we modified our test suite to use check-state with
record-state in firewall rules. Now, we want to extend our tests to
also cover keep-state functionality.

To achieve this, we need to duplicate the existing tests with slight
modifications to the firewall ruleset to use keep-state instead of check-state
with record-state. But since each other config piece remains the same we just
expand the autotest.yaml file

There might be a better way to handle this though, but current autotest.cpp
does not support explicitly specifying the firewall.txt file.
Thus reusing the "reload" function, as the firewall.txt paths are defined
within the controlplane.conf files.