Add VariantTraitMap to track action indices with specific traits
Introduce the `VariantTraitMap`, a generic helper that maps types within a
`std::variant` to a value (e.g., `std::optional<size_t>`) based on a given trait.

Why this is needed:

Some actions, (those with `MAX_COUNT=1`), only need to be executed once or
only the first/last occurrence of them matters. For example:

- CheckStateAction:
    Only the first occurrence of this action is meaningful.
- StateTimeoutAction:
    Only the last occurrence of this action should be stored and processed,
    as it provides the most accurate or latest information.

To optimize contolplane, we can store only one of those actions with their index.
This way, on "compile" step in value_t we can do something with this information
to increase contolplane performance. For example, compeletly ditch the
StateTimeoutAction and just add the timeout to the FlowAction as a field.

Therefore, to avoid writing boilerplate code like
`std::optional<size_t> check_state_index, state_timeout_index`, we can
use VariantTraitMap to directly map Actions with MAX_COUNT=1 to their index.