Add common/utils.h with small utility functions
This commit introduces a new common/utils.h file to store small utility
functions, as such a file didn't previously exist.

It also adds the always_false trait, which is a trait that evaluates to false
for any type. This trait is useful for triggering static assertions in order
to identify the exact type that caused a compilation error.