Fill state_timeout only on creation
acl_touch_state is used only in acl_create_state and acl_checkstate.
The former creates the state, while the latter attempts to match a packet
with it. During state creation, we need to fill the `value` with the
appropriate timeout, which is initialized to zero by default. However,
in acl_checkstate, we still need to "touch" the state, ensuring that the
last_seen field is updated with the current time.

The issue with the current code is that it unnecessarily fills the
state_timeout value every time, even though this value doesn't change
once set. It is defined during state creation and later used to check
for timeouts, but we don't modify it afterward. Recalculating it each
time adds extra complexity/branches to the dataplane without any benefit.
This commit removes that behavior.