Mark unused structured bindings with YANET_GCC_BUG_UNUSED instead of void cast
GCC 7.5 version which we have to support for now due to Ubuntu 18 have a bug
(https://gcc.gnu.org/bugzilla/show_bug.cgi?id=81767) where unused variables in
structured bindings are incorrectly flagged as warnings, even though they
"can't" be used, cause they are not a real variables.

Current code uses `(void)var;` to suppress these warnings, but this approach
is ugly and harder to grep.

This commit removes `(void)var;` statements and replaces it with `YANET_GCC_BUG_UNUSED`

Once we move to a newer GCC version where this bug is fixed,
we can easily remove the that macro as it is much easier to grep and remove
than those void(var) statements