Add dummy hitcount rule
This commit introduces parsing for the hitcount rule.
The tests and the actual semantics will be implemented later.

The hitcount action is designed to count the number of times a rule is hit.
It will be introduced in a ruleset as a duplecate of regular rule,
but its action is hitcount instead of a terminating action.

The hitcount values will be aggregated and returned upon request.

We use "hitcount" instead of just "count" to not interfere with IPFW
"count" rule.