Added the ability to set route tunnel routes in controlplane.conf
In controlplane.conf, it was possible to set static routes for route. Added
the ability to also set routes for route tunnel, to do this, set the route:
	"tunnel": true
One can also set the path_information and label parameters. If these parameters
are not set, the default values will be used.