transport_table: Use fast flat hashmap in remap_group_ids
Previously the remap mapping was stored in a vector that was resized to group_id,
which could grow into the 800k–820k range. Since most entries were zero,
the vector was huge and inefficient. This change replaces the vector with a
cache-friendly emhash7::HashMap

Results in ~1gb saved memory and compile step being four times faster