transport_table: Change group_id_filter_ids map-to-set to only set
The one and only usage of group_id_filter_ids is to obtain it's size.
We're not even interested in the fact that this is a map that maps
unique groups to unique filters.

So we can just use unordered_set to get the size we need in the end