Caluclate unique group_ids size in-place in ACL_DEBUG mode
We got rid of "bitset" which was used not only for filling
group_ids per-filter, but also in ACL_DEBUG mode it allowed
us to get unique group_ids amount.

We can just calculate this by shoving all group_ids from
accumulated sets into one big set and getting it's size.

Another possible approach is to shove all values to a vector,
sort it and erase duplicates with std::unique. This could help with
very large amounts due to better cache utilization, but in practice
it's negligible, especially in debug mode.