transport: Remove unnecessary set copies in 'distribute' step
To find the best layer, the old approach looked for an existing layer
where adding new filter IDs doesn't increase the total count.
It did so by copying set, merging it with the current one and comparing
it's size. The only thing that we needed from that resulting set is it's
size.

We can acheive the same in a more efficient way by checking if a perfect match
exists first without creating a copy of a set. Then, if the current
layer is not the perfect match, we can simply calculate the "new" set size
instead of actually creating it.