acl_network_table: Remove bitmask, store sets directly where we need them
This is a direct throwback to the same previous commit about acl_transport_table.

Just like before, `bitmask` helps us fill a data structure
that allows us to get unique group_ids **per-filter**.

We fill bitmask with group_ids that we get from the 2-dimensional
table per-filter. Since it's a set, we get **unique** group_ids per filter.
But we don't need to fill that `bitmask` structure and then iterate
over it, copying it's contents to the vector, and clean it up each time.
We can just define filter_id_group_ids as vector of sets
instead of vector of vectors.

Speeds up "populate" step by 55%