Add flat hashmap 'emhash' as a subproject
Also the custom_hash_structures option is added to optionally replace
std::unordered_map with emhash7::HashMap. This is because
unordered_map is sometimes slow due to being a node-based map,
which provides benefits like iterator non-invalidation guarantees.
However, in the ACL maps are primarily used for filling and iterating.
A flat map structure like emhash7::HashMap is more desirable for performance.

Reasoning behind choosing this particular flat hashmap implementation
is in the next commit.