transport_table: make group_id_filter_ids use vector instead of a set
In practice, a group_id corresponds to a low amount of unique
filter_ids, about a hundered. But the amount of times we try to insert a
filter and it's already presented is huge (up to a half a million) due
to necessary loops over 6-dimensional table.

Therefore to improve performance we can just use linear find over that
small amount. We do not resize those vector to filters.size() as this
value is drastically higher than a number of unique filters per group_id