Add test to verify current keep-state behavior
Introduce a new test to verify the current behavior of the keep-state.
Currently, when a rule is marked with keep-state, the parser implicitly
adds a check-state rule before it. This means that every keep-state rule
results in two rules: one check-state rule and the original rule
with record-state.

We want to optimize that.