Make dataplane/sharedmemory.cpp actually manage dump rings
This file was not related to sharedmemory management at all.
The naming was bad, since this code manages dump rings. Renamed it
to dump_rings then.

Dump ring is represented by either RingRaw or RingPcap structure.
They implement common interface defined in RingBase, which is, write a
packet and get a packet. RingPcap implementation is incomplete by now
and will be added in later commits.

Also there's a "GetCapacity" function that helps on ring creation a lot