Flush DumpRing buffer before reading packets in autotests
When running YANET with autotests, we have two processes - the dataplane and
an autotest. The dataplane allocates shared memory and creates DumpRing objects
that use this memory to store captured packets. The autotest creates
it's own DumpRing object using the same shared memory and uses them to
obtain stored packets.
There are two types of these DumpRings: the original RawRing and a RingPcap.

Since a RingPcap format uses libpcap to write packets to a FILE*, it
uses buffering. In order to make sure we have all the packets in the
shared memory, first we need to flush the buffer from the DumpRing
object in dataplane.