Refactor initEal function
Made buffer manage memory automatically. The fixed-size char buffer[8192]
and manual position tracking is error prone.

Remove insert_eal_arg macro as it obscures logic and lacks type safety.
Now we just store arguments in a vector and convert it to C-style array in the end.

std::bitset<64> and its to_ullong() method are remade to allow the use of more than 64 cores.