Prevent SHM ring buffer corruption when reader is too slow
When `yanet-cli tcpdump follow` is used, it acts as a consumer for one or more
shared memory ring buffers being filled by `dataplane` producer threads.
A bug existed where the consumer could read corrupted data if it fell too far
behind the producer.

Each producer (dataplane worker) uses two `uint64_t` atomic counters to manage the ring buffer:
- `writer_pos`: The absolute offset where the writer is writing.
- `reader_pos`: The absolute offset where the reader is reading.

The system worked correctly as long as `writer_pos - reader_pos <= ring_buffer_size`.
However, `yanet-cli` (which writes to stdout) is significantly slower than `dataplane`.
That means that the producer could "lap" the consumer, wrapping around
the entire buffer and overwriting data that the consumer hadn't read yet.

When this happened, the consumer would see a large gap between `writer_pos`
and `reader_pos`, but would be unaware that the data at its current `reader_pos`
was now invalid. It would attempt to interpret this new data as an old packet,
leading to corrupted pcap output and errors like `[Invalid header: len==0]`.

Example:
1.  `yanet-cli` starts and attaches to the shared memory.
2.  A large burst of packets hits the dataplane. The producer thread fills the
    entire ring buffer and wraps around in milliseconds.
3.  `yanet-cli`'s polling loop finally gets scheduled. It sees that `writer_pos`
    has advanced by an amount greater than the ring buffer's size.
4.  Without a check, it would start reading from `reader_pos`,
    which now points to the middle of a much newer packet, which led to data corruption.

Now when a lap is detected, we print a warning to `stderr` and discard all packets in the buffer.

In real-life YANET usage we process about 2 million packets per second. We cannot
write all of them to stdout even if we really want to, because `dataplane` is
that much faster. Discarding those packets is our only option.