Implement chunked allocation for dynamic hashtable
Previously, the hashtable was allocated as a single, contiguous block of memory.
This approach failed when requesting gigabytes of memory, as a sufficiently large
contiguous block could not be found, even when total available memory was high.

This refactoring introduces changes for `hashtable_mod_id32_dynamic` and related `updater_hashtable_mod_id32`:

- The internal storage is changed from a flexible array member to a pointer to an array of pointers
- A new `get_pair_ptr()` method translates a logical table index into a physical memory location
  by calculating the correct chunk and the offset within it.

The updater now manages the following allocation process:
- allocates the small, fixed-size hashtable header
- allocates the array of pointers for the chunks
- allocates each data chunk individually in a loop